+++ src/dslice_eqp_k.cpp
@@ -4,13 +4,13 @@ using namespace Rcpp;
 // [[Rcpp::export]]
 Rcpp::List dslice_eqp_k(Rcpp::NumericVector x, int dim, double lambda)
 {
-  int len = x.size();
-	double lpd = -lambda * log(len);
+	int len = x.size();
+	double lpd = -lambda * log((double)len);
 	const double epsilon = 1e-6;
 
-	int baselen = sqrt(len);
-  //  at most "baselen+1" groups with each group has size "baselen" at least
-  Rcpp::NumericVector divsch(baselen+3);
+	int baselen = (int)sqrt((double)len);
+	//  at most "baselen+1" groups with each group has size "baselen" at least
+	Rcpp::NumericVector divsch(baselen+3);
 	divsch[0] = 0;
 	int flagl = baselen;
 	int ngrp = 0;
@@ -23,7 +23,7 @@ Rcpp::List dslice_eqp_k(Rcpp::NumericVector x, int dim, double lambda)
 		}
 	}
 	divsch[++ngrp] = len;
-  Rcpp::NumericMatrix ctab(ngrp+1, dim);
+	Rcpp::NumericMatrix ctab(ngrp+1, dim);
 	for(int k = 1; k < ngrp+1; ++k){
 		for(int j = 0; j < dim; ++j){
 			ctab(k, j) += ctab(k-1, j);
@@ -32,8 +32,8 @@ Rcpp::List dslice_eqp_k(Rcpp::NumericVector x, int dim, double lambda)
 			ctab(k, x[j]) += 1;
 		}
 	}
-  Rcpp::NumericVector score(ngrp+1);
-  Rcpp::IntegerVector idx(ngrp+1);
+	Rcpp::NumericVector score(ngrp+1);
+	Rcpp::IntegerVector idx(ngrp+1);
 	for(int k = 0; k < ngrp+1; ++k){
 		score[k] = 0;
 		idx[k] = -1;
@@ -48,7 +48,7 @@ Rcpp::List dslice_eqp_k(Rcpp::NumericVector x, int dim, double lambda)
 		for(int u = 0; u < dim; ++u){
 			tc += ctab(i, u);
 		}
-  	for(int u = 0; u < dim; ++u){
+		for(int u = 0; u < dim; ++u){
 			counts[u] = ctab(i, u);
 			if(counts[u] > epsilon){
 				cutsc += counts[u] * log(counts[u] / tc);
@@ -56,7 +56,7 @@ Rcpp::List dslice_eqp_k(Rcpp::NumericVector x, int dim, double lambda)
 		}
 		cutpos = 0;
 		//  j = 0 end
-  	for(int j = 1; j < i; ++j){
+		for(int j = 1; j < i; ++j){
 			tpcut = lpd + score[j];
 			tc = 0;
 			for(int u = 0; u < dim; ++u){
@@ -84,35 +84,35 @@ Rcpp::List dslice_eqp_k(Rcpp::NumericVector x, int dim, double lambda)
 		slicenum++;
 	}
   
-  Rcpp::IntegerMatrix slices(slicenum, dim+1);
-  Rcpp::CharacterVector rownames(slicenum);
-  Rcpp::CharacterVector colnames(dim+1);
-  std::string s = "s";
-  std::string instr;
-  for(int i = 0; i < slicenum; ++i){
-    std::stringstream ss;
-    ss << i + 1;
-    ss >> instr;
-    rownames[i] = s + instr;
-  }
-  for(int j = 0; j < dim; ++j){
-    std::stringstream ss;
-    ss << j;
-    ss >> instr;
-    colnames[j] = instr;
-  }
-  colnames[dim] = "total";
-  Rcpp::List dimnames = Rcpp::List::create(rownames, colnames);
-  slices.attr("dimnames") = dimnames;
-  
-  Rcpp::IntegerVector spos(slicenum+1);
-  flag = ngrp;
+	Rcpp::IntegerMatrix slices(slicenum, dim+1);
+	Rcpp::CharacterVector rownames(slicenum);
+	Rcpp::CharacterVector colnames(dim+1);
+	std::string s = "s";
+	std::string instr;
+	for(int i = 0; i < slicenum; ++i){
+		std::stringstream ss;
+		ss << i + 1;
+		ss >> instr;
+		rownames[i] = s + instr;
+	}
+	for(int j = 0; j < dim; ++j){
+		std::stringstream ss;
+		ss << j;
+		ss >> instr;
+		colnames[j] = instr;
+	}
+	colnames[dim] = "total";
+	Rcpp::List dimnames = Rcpp::List::create(rownames, colnames);
+	slices.attr("dimnames") = dimnames;
+
+	Rcpp::IntegerVector spos(slicenum+1);
+	flag = ngrp;
 	for(int i = slicenum; i > -1; --i){
 		spos[i] = flag;
 		flag = idx[flag];
 	}
 	spos[0] = 0;
-  for(int i = 0; i < slicenum; ++i){
+	for(int i = 0; i < slicenum; ++i){
 		for(int j = 0; j < dim; ++j){
 			slices(i, j) = ctab(spos[i+1], j) - ctab(spos[i], j);
 			slices(i, dim) += slices(i, j);
@@ -120,11 +120,11 @@ Rcpp::List dslice_eqp_k(Rcpp::NumericVector x, int dim, double lambda)
 	}
 	double mlik = score[ngrp] - lpd;
 	//  substract null log-likelihood (assume one slice, i.e., no cut)
-  for(int u = 0; u < dim; ++u){
+	for(int u = 0; u < dim; ++u){
 		if(ctab(ngrp, u) > epsilon){
 			mlik -= ctab(ngrp, u) * log(ctab(ngrp, u) / len);
 		}
 	}
 	Rcpp::List slicing_res = Rcpp::List::create(Rcpp::Named("dsval")=mlik, Rcpp::Named("slices")=slices);
-  return slicing_res;
+	return slicing_res;
 }
